requirements.txt
streamlit==0.68.0
pandas==1.1.3
plotly==4.8.0
matplotlib==3.2.2
Pillow==8.0.0
